task1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	int size = 8 * sizeof(unsigned long int);
+	unsigned long int size = 8 * sizeof(unsigned long int);
 
 	while (size)
 	{
